fix: make it work as GiHub dependency when using Vite
When using `@libp2p/webrtc` directly with pointing to a GitHub repository,
it won't build with Vite, due to the defined `exports` in the package.json.
One way to make it work is adding the build step as `prepare` script.

Steps to reproduce the issue:

    npm create vite@latest -- --template vanilla webrtcasgithubdep
    cd webrtcasgithubdep
    npm install https://github.com/libp2p/js-libp2p-webrtc
    echo "import { webRTC } from '@libp2p/webrtc'" > main.js
    npm run build

Errors with:

    [commonjs--resolver] Failed to resolve entry for package "@libp2p/webrtc". The package may have incorrect main/module/exports specified in its package.json.
error during build

With the prepare script it builds without errors.